fix: Add multiple mentions (#939)
* fix: Stop matching mentions at next @

* refactor: Remove unneccessary nested template string

* Fix: avoid trim when checking for tokens

This is a **hack** to avoid an issue with the regex we use to match
mentions expecting a user to have two names, to determine that a the
input is/contains a mention.

* fix: get start of new token based on caret placement

The regex used for this so far matches an `@` and then any following
charactes except whitespace and `@`, in practice one word. Therefore,
we don't actually know that we search for the right trigger.

This looks for the previous trigger backwards in the string to be
modified, limited at where the current caret is.